# Conflicts:
#	docs/JPO_ODE_UserGuide.docx
#	jpo-ode-plugins/src/main/java/us/dot/its/jpo/ode/plugin/builders/timstorage/DataFrames.java
#	jpo-ode-plugins/src/main/java/us/dot/its/jpo/ode/plugin/builders/timstorage/Tim.java
#	jpo-ode-plugins/src/main/java/us/dot/its/jpo/ode/plugin/j2735/builders/HeadingBuilder.java
#	jpo-ode-plugins/src/main/java/us/dot/its/jpo/ode/plugin/j2735/builders/TravelerMessageFromHumanToAsnConverter.java
#	jpo-ode-plugins/src/main/java/us/dot/its/jpo/ode/plugin/j2735/builders/timstorage/DataFrame.java
#	jpo-ode-plugins/src/main/java/us/dot/its/jpo/ode/plugin/j2735/builders/timstorage/DataFrames.java
#	jpo-ode-svcs/src/main/java/us/dot/its/jpo/ode/OdeProperties.java
#	jpo-ode-svcs/src/main/java/us/dot/its/jpo/ode/coder/TimDecoderHelper.java
#	jpo-ode-svcs/src/main/java/us/dot/its/jpo/ode/services/asn1/Asn1EncodedDataRouter.java
#	jpo-ode-svcs/src/main/java/us/dot/its/jpo/ode/traveler/TimController.java
#	jpo-ode-svcs/src/test/java/us/dot/its/jpo/ode/traveler/AsdMessageTest.java
#	jpo-ode-svcs/src/test/java/us/dot/its/jpo/ode/traveler/TimControllerTest.java

+++ jpo-ode-plugins/src/main/java/us/dot/its/jpo/ode/plugin/j2735/builders/OffsetLLB24Builder.java
@@ -8,7 +8,7 @@ public class OffsetLLB24Builder {
       throw new UnsupportedOperationException();
    }
 
-   public static int offsetLLB24(BigDecimal offset) {
-      return offset.scaleByPowerOfTen(7).intValue();
+   public static Long offsetLLB24(BigDecimal offset) {
+      return offset.scaleByPowerOfTen(7).longValue();
    }
 }
